Fix populateCgroupEntryListV2
cgroupFile - /proc/<PID>/cgroup

There can be multiple lines in the cgroupFile. This happens when
cgroups v1 controllers, which do not have a cgroups v2 equivalent,
are enabled in a cgroups v2 environment.

In this case, all lines in the cgroupFile need to be parsed to
locate the line with the '0::<cgroup>' format.

Use sizeof to derive the length of the char array.

Related: #6935

Signed-off-by: Babneet Singh <sbabneet@ca.ibm.com>